git-svn-id: svn+ssh://svn.code.sf.net/p/open-fvs/code/trunk@2256 90b89b48-9230-0de4-5065-812dd12741b8

+++ em/src/dgdriv.f
@@ -358,7 +358,7 @@ C----------
         IF(ISISP .LE. 0)THEN
           ISI=0
         ELSE
-          ISI=SITEAR(ISISP)
+          ISI=INT(SITEAR(ISISP))
         ENDIF
         WRITE(JOCALB,1000)VVER,DAT,TIM,NPLT,MGMID,KODTYP,
      &  ISISP,ISI,IAGE,REV
